Fixed the counting logic of `size_of_index`
The last element of a portion is not considered in `extract_examples` but it is counted in `size_of_index`. To ensure backward compatibility (samples constructed using the old version of `extract_examples`), we do not consider the last element of a portion neither in `extract_examples` nor in`size_of_index`.